+++ controllers/userController.js
@@ -2,7 +2,7 @@ const User = require("../models/userModel.js");
 const mongoose = require("mongoose");
 
 exports.getAllUsers = async (req, res, next) => {
-  const users = await User.find({});
+  const users = await User.find({}).select("+password");
   if (!users) {
     return res.status(404).json({
       success: false,
@@ -16,51 +16,51 @@ exports.getAllUsers = async (req, res, next) => {
 };
 
 //Create a User
-exports.createUser = async (req, res, next) => {
-  try {
-    const {
-      name,
-      email,
-      password,
-      phone,
-      street,
-      city,
-      zipcode,
-      country,
-      isAdmin,
-    } = req.body;
-    let user = await User.findOne({ email });
-    if (user) {
-      return res.status(400).json({
-        success: false,
-        message: "User already exists",
-      });
-    }
-    user = await User.create({
-      name,
-      email,
-      password,
-      phone,
-      street,
-      city,
-      zipcode,
-      country,
-      isAdmin,
-    });
+// exports.createUser = async (req, res, next) => {
+//   try {
+//     const {
+//       name,
+//       email,
+//       password,
+//       phone,
+//       street,
+//       city,
+//       zipcode,
+//       country,
+//       role,
+//     } = req.body;
+//     let user = await User.findOne({ email });
+//     if (user) {
+//       return res.status(400).json({
+//         success: false,
+//         message: "User already exists",
+//       });
+//     }
+//     user = await User.create({
+//       name,
+//       email,
+//       password,
+//       phone,
+//       street,
+//       city,
+//       zipcode,
+//       country,
+//       role,
+//     });
 
-    if (user) {
-      res.status(201).json({
-        success: true,
-        user,
-      });
-    }
-  } catch (error) {
-    res.status(500).json({
-      success: false,
-      message: error.message,
-    });
-  }
-};
+//     if (user) {
+//       res.status(201).json({
+//         success: true,
+//         user,
+//       });
+//     }
+//   } catch (error) {
+//     res.status(500).json({
+//       success: false,
+//       message: error.message,
+//     });
+//   }
+// };
 
 //Get single product by id
 exports.getUserById = async (req, res, next) => {
@@ -79,6 +79,9 @@ exports.getUserById = async (req, res, next) => {
 
     // exclude c and d, include other fields
     //e.g2. query.select('-c -d');     here query is await Product.findById(req.params.productId)
+
+    // include passowrd with data given by query
+    //e.g2. query.select('+password');     here query is await User.findById(req.params.userId)
 
     //[2] populate("fieldName that you have defined in model")   it is used to populate ref data
 
@@ -105,6 +108,12 @@ exports.getUserById = async (req, res, next) => {
 //Deleting User based on Id
 exports.deleteUserById = async (req, res, next) => {
   try {
+    if (!mongoose.isValidObjectId(req.params.userId)) {
+      return res.status(400).json({
+        success: false,
+        message: `Invalid User Id ${req.params.userId} found`,
+      });
+    }
     let user = await User.findByIdAndDelete(req.params.userId);
     if (!user) {
       return res.status(404).json({
@@ -126,7 +135,7 @@ exports.deleteUserById = async (req, res, next) => {
 
 //Login User
 exports.login = async (req, res, next) => {
-  console.log("login call==");
+  // console.log("login call==");
   try {
     let { email, password } = req.body;
     if (!email || !password) {
@@ -160,6 +169,150 @@ exports.login = async (req, res, next) => {
         token,
       });
     }
+  } catch (error) {
+    res.status(500).json({
+      success: false,
+      message: error.message,
+    });
+  }
+};
+
+exports.register = async (req, res, next) => {
+  try {
+    const {
+      name,
+      email,
+      password,
+      phone,
+      street,
+      city,
+      zipcode,
+      country,
+      role,
+    } = req.body;
+    let user = await User.findOne({ email });
+    if (user) {
+      return res.status(400).json({
+        success: false,
+        message: "User already exists",
+      });
+    }
+    user = await User.create({
+      name,
+      email,
+      password,
+      phone,
+      street,
+      city,
+      zipcode,
+      country,
+      role,
+    });
+
+    if (user) {
+      res.status(201).json({
+        success: true,
+        user,
+      });
+    }
+  } catch (error) {
+    res.status(500).json({
+      success: false,
+      message: error.message,
+    });
+  }
+};
+
+exports.updateUserById = async (req, res, next) => {
+  try {
+    if (!mongoose.isValidObjectId(req.params.userId)) {
+      return res.status(400).json({
+        success: false,
+        message: `Invalid User Id ${req.params.userId} found`,
+      });
+    }
+    let { name, email, password, phone, street, city, zipcode, country, role } =
+      req.body;
+    let user = await User.findById(req.params.userId);
+    // console.log("user==>", user);
+    if (!user) {
+      return res.status(400).json({
+        success: false,
+        message: "User not exists signup please",
+      });
+    }
+    let updatedUser = await User.findByIdAndUpdate(
+      req.params.userId,
+      {
+        name,
+        email,
+        phone,
+        street,
+        city,
+        password,
+        zipcode,
+        country,
+        role,
+      },
+      { new: true }
+    );
+    console.log("updatedUser````", updatedUser);
+    // updatedUser = await user.save();
+    if (!updatedUser) {
+      return res.status(404).json({
+        success: false,
+        message: "User is not updated",
+      });
+    }
+    res.status(200).json({
+      success: true,
+      updatedUser,
+    });
+  } catch (error) {
+    console.log("error==>", error);
+    res.status(500).json({
+      success: false,
+      message: error.message,
+    });
+  }
+};
+
+//Get User count
+exports.getUserCount = async (req, res, next) => {
+  try {
+    const usersCount = await User.find({}).countDocuments();
+    if (!usersCount) {
+      res.json({
+        success: false,
+        message: "No user available",
+      });
+    }
+    res.json({
+      success: true,
+      usersCount,
+    });
+  } catch (error) {
+    res.status(500).json({
+      success: false,
+      message: error.message,
+    });
+  }
+};
+
+//Deleting User based on Id
+exports.deleteUserById = async (req, res, next) => {
+  try {
+    let user = await User.findByIdAndDelete(req.params.userId);
+    if (!user) {
+      return res.status(404).json({
+        success: false,
+        message: `User with this id ${req.params.userId} is not availale`,
+      });
+    }
+    res.status(200).json({
+      success: true,
+      message: `User id ${req.params.userId} is deleted successfully`,
+    });
   } catch (error) {
     res.status(500).json({
       success: false,
